put all plugin related stuff in one file plugin_manager
