+++ metagov/metagov/plugins/github/models.py
@@ -1,8 +1,7 @@
 import requests, json, logging
 from collections import Counter
 
-import metagov.core.plugin_decorators as Registry
-from metagov.core.plugin_constants import Parameters
+from metagov.core.plugin_manager import Registry, Parameters, VotingStandard
 from metagov.core.models import Plugin, GovernanceProcess, ProcessStatus, AuthType
 from metagov.core.errors import PluginErrorInternal
 import metagov.plugins.github.schemas as Schemas
