+++ metagov/metagov/plugins/loomio/models.py
@@ -1,9 +1,8 @@
 import json
 import logging
 
-import metagov.core.plugin_decorators as Registry
+from metagov.core.plugin_manager import Registry, Parameters, VotingStandard
 import metagov.plugins.loomio.schemas as Schemas
-from metagov.core.plugin_constants import Parameters
 import requests
 from metagov.core.errors import PluginErrorInternal
 from metagov.core.models import GovernanceProcess, Plugin, ProcessStatus, AuthType
